Determine the demographic breakout of CiHs by High School

Determine the demographic breakout of CiHs by High School

ciHS_demographics(data, high_school_name = "")

Arguments

data
The DW STUDENT table saved as an R object.

high_school_name
Provide the entire or partial High School name.

Value

A tibble with number of students per high school broken out by race/ethnicity.

Examples

library(tibble)
library(magrittr)

# The data will look something like this.
df <- tibble(
   sid = c("1" , "2"),
   year = c("B90", "B90"),
   hi_schl = c("178", "179"),
   race_ethnic_code = c("1", "6"),
   dual_enroll = c("2", "1")
   )

 df

#> # A tibble: 2 x 5
#>   sid   year  hi_schl race_ethnic_code dual_enroll
#>   <chr> <chr> <chr>   <chr>            <chr>      
#> 1 1     B90   178     1                2          
#> 2 2     B90   179     6                1          

# Providing no high school will return all the values
 df %>%
    ciHS_demographics()

#> # A tibble: 1 x 4
#>   year  race_ethnic_code high_school_name        n
#>   <chr> <chr>            <chr>               <int>
#> 1 B90   A/W              Moses Lake Hgh Schl     1

# Using either the full name or paritial name will suffice if spelled correctly.
 df %>%
    ciHS_demographics(high_school_name = "Moses Lake")

#> # A tibble: 1 x 4
#>   year  race_ethnic_code high_school_name        n
#>   <chr> <chr>            <chr>               <int>
#> 1 B90   A/W              Moses Lake Hgh Schl     1