Clean DW Transcript Table

Used to clean the transcript table by removing non-credit course, community courses, or any other not applicable courses.

clean_dw_transcript(data)

Arguments

data
Name of R object that contains the TRANSCRIPT data.

Value

A dataframe with a cleaned transcript table

Examples

library(tibble)
library(magrittr)

df <- tibble(
        dept_div = c("NSO", "DVS", "CHEM&", "ENGL", "BDC"),
        course_num = c("100", "098", "121", "099", "121"),
        cr = c(NA, NA, 5, 5, 2),
        item = c("XOXO", "1", "2", "3", "4"),
        sect = c("A#P", "HY", "OL1", NA, "OL3"),
        class_size = c(500, 20, 25, 25, 25)
        )

 df

#> # A tibble: 5 x 6
#>   dept_div course_num    cr item  sect  class_size
#>   <chr>    <chr>      <dbl> <chr> <chr>      <dbl>
#> 1 NSO      100           NA XOXO  A#P          500
#> 2 DVS      098           NA 1     HY            20
#> 3 CHEM&    121            5 2     OL1           25
#> 4 ENGL     099            5 3     NA            25
#> 5 BDC      121            2 4     OL3           25

# Piping clean_dw_transcripts will return those courses that should
# be excluded from course success.

 df %>%
   clean_dw_transcript()

#> # A tibble: 3 x 6
#>   dept_div course_num    cr item  sect  class_size
#>   <chr>    <chr>      <dbl> <chr> <chr>      <dbl>
#> 1 CHEM&    121            5 2     OL1           25
#> 2 ENGL     099            5 3     NA            25
#> 3 BDC      121            2 4     OL3           25